feat(builder-shared): more plugin sink in shared
