+++ .changeset/four-bags-perform.md
@@ -1,0 +1,8 @@
+---
+'@modern-js/builder-rspack-provider': minor
+'@modern-js/builder-shared': minor
+'@modern-js/builder-webpack-provider': minor
+---
+
+feat: more plugin rebase in `builder-shared`
+feat: 更多 plugin 下沉到 `builder-shared`
