--- packages/builder/builder-webpack-provider/static/template.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <%= meta %>
-  <title><%= title %></title>
-</head>
-<body>
-  <div id="<%= mountId %>"></div>
-</body>
-</html>
